+++ db/migrate/20200518084242_add_popularity_ratio_to_service.rb
@@ -1,0 +1,5 @@
+class AddPopularityRatioToService < ActiveRecord::Migration[6.0]
+  def change
+    add_column :services, :popularity_ratio, :float
+  end
+end